+++ packages/orq-rc/RELEASES.md
@@ -4899,3 +4899,13 @@ Based on:
 - [typescript v4.0.0-rc.21] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.21 - packages/orq-rc
+
+## 2025-11-11 07:18:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.653.0 (2.748.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.22] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.22 - packages/orq-rc